Use pnp instead of node-modules for Yarn nodeLinker
Unrecognized or legacy configuration settings found: node-linker - run "yarn config -v" to see the list of settings supported in Yarn

Set nodeLinker which defines what linker should be used for installing Node packages (useful to enable the node-modules plugin), one of: pnp, pnpm and node-modules.

node-linker setting was completely wrong and trying pnp: points to cache where the file actually is. It uses a file called pnp.jswhich acts as a map to where in the cache these are located.
https://medium.com/@Andrew_Mc/yarn-pnp-npm-tink-and-pnpm-oh-my-720d02221b4b